+++ mne/mixed_norm/optim.py
@@ -3,7 +3,7 @@
 # License: Simplified BSD
 
 import warnings
-from math import sqrt
+from math import sqrt, ceil
 import numpy as np
 from scipy import linalg
 
@@ -12,6 +12,7 @@ logger = logging.getLogger('mne')
 
 from .debiasing import compute_bias
 from .. import verbose
+from ..time_frequency.stft import stft_norm2, stft, istft
 
 
 def groups_norm2(A, n_orient):
@@ -38,10 +39,13 @@ def norm_l21(A, n_orient, copy=True):
     return np.sum(np.sqrt(groups_norm2(A, n_orient)))
 
 
-def prox_l21(Y, alpha, n_orient):
+def prox_l21(Y, alpha, n_orient, shape=None, is_stft=False):
     """proximity operator for l21 norm
 
-    (L2 over columns and L1 over rows => groups contain n_orient rows)
+    L2 over columns and L1 over rows => groups contain n_orient rows.
+
+    It can eventually take into account the negative frequencies
+    when a complex value is passed and is_stft=True.
 
     Example
     -------
@@ -60,11 +64,17 @@ def prox_l21(Y, alpha, n_orient):
     [ True  True False False]
     """
     if len(Y) == 0:
-        return np.zeros((0, Y.shape[1]), dtype=Y.dtype), \
-                    np.zeros((0,), dtype=np.bool)
+        return np.zeros_like(Y), np.zeros((0,), dtype=np.bool)
+    if shape is not None:
+        shape_init = Y.shape
+        Y = Y.reshape(*shape)
     n_positions = Y.shape[0] // n_orient
-    rows_norm = np.sqrt(np.sum((np.abs(Y) ** 2).reshape(n_positions, -1),
-                                axis=1))
+
+    if is_stft:
+        rows_norm = np.sqrt(stft_norm2(Y).reshape(n_positions, -1).sum(axis=1))
+    else:
+        rows_norm = np.sqrt(np.sum((np.abs(Y) ** 2).reshape(n_positions, -1),
+                                    axis=1))
     # Ensure shrink is >= 0 while avoiding any division by zero
     shrink = np.maximum(1.0 - alpha / np.maximum(rows_norm, alpha), 0.0)
     active_set = shrink > 0.0
@@ -72,7 +82,11 @@ def prox_l21(Y, alpha, n_orient):
         active_set = np.tile(active_set[:, None], [1, n_orient]).ravel()
         shrink = np.tile(shrink[:, None], [1, n_orient]).ravel()
     Y = Y[active_set]
-    Y *= shrink[active_set][:, np.newaxis]
+    if shape is None:
+        Y *= shrink[active_set][:, np.newaxis]
+    else:
+        Y *= shrink[active_set][:, np.newaxis, np.newaxis]
+        Y = Y.reshape(-1, *shape_init[1:])
     return Y, active_set
 
 
@@ -204,7 +218,7 @@ def _mixed_norm_solver_prox(M, G, alpha, maxit=200, tol=1e-8, verbose=None,
 
         t0 = t
         t = 0.5 * (1.0 + sqrt(1.0 + 4.0 * t ** 2))
-        Y[:] = 0.0
+        Y.fill(0.0)
         dt = ((t0 - 1.0) / t)
         Y[active_set] = (1.0 + dt) * X
         Y[active_set_0] -= dt * X0
@@ -264,6 +278,9 @@ def mixed_norm_solver(M, G, alpha, maxit=3000, tol=1e-8, verbose=None,
         The data
     G : array
         The forward operator
+    alpha : float
+        The regularization parameter. It should be between 0 and 100.
+        A value of 100 will lead to an empty active set (no active source).
     maxit : int
         The number of iterations
     tol : float
@@ -281,12 +298,12 @@ def mixed_norm_solver(M, G, alpha, maxit=3000, tol=1e-8, verbose=None,
 
     Returns
     -------
-    X: array
-        The source estimates
-    active_set: array
-        The mask of active sources
-    E: array
-        The cost function over the iterations
+    X : array
+        The source estimates.
+    active_set : array
+        The mask of active sources.
+    E : list
+        The value of the objective function over the iterations.
     """
     n_dipoles = G.shape[1]
     n_positions = n_dipoles // n_orient
@@ -372,6 +389,239 @@ def mixed_norm_solver(M, G, alpha, maxit=3000, tol=1e-8, verbose=None,
     else:
         X, active_set, E = l21_solver(M, G, alpha, maxit=maxit,
                                       tol=tol, n_orient=n_orient)
+
+    if (active_set.sum() > 0) and debias:
+        bias = compute_bias(M, G[:, active_set], X, n_orient=n_orient)
+        X *= bias[:, np.newaxis]
+
+    return X, active_set, E
+
+
+###############################################################################
+# TF-MxNE
+
+@verbose
+def tf_lipschitz_constant(M, G, phi, phiT, tol=1e-3, verbose=None):
+    """Compute lipschitz constant for FISTA
+
+    It uses a power iteration method.
+    """
+    n_times = M.shape[1]
+    n_points = G.shape[1]
+    iv = np.ones((n_points, n_times), dtype=np.float)
+    v = phi(iv)
+    L = 1e100
+    for it in range(100):
+        L_old = L
+        logger.info('Lipschitz estimation: iteration = %d' % it)
+        iv = np.real(phiT(v))
+        Gv = np.dot(G, iv)
+        GtGv = np.dot(G.T, Gv)
+        w = phi(GtGv)
+        L = np.max(np.abs(w))  # l_inf norm
+        v = w / L
+        if abs((L - L_old) / L_old) < tol:
+            break
+    return L
+
+
+def safe_max_abs(A, ia):
+    """Compute np.max(np.abs(A[ia])) possible with empty A"""
+    if np.sum(ia):  # ia is not empty
+        return np.max(np.abs(A[ia]))
+    else:
+        return 0.
+
+
+def safe_max_abs_diff(A, ia, B, ib):
+    """Compute np.max(np.abs(A)) possible with empty A"""
+    A = A[ia] if np.sum(ia) else 0.0
+    B = B[ib] if np.sum(ia) else 0.0
+    return np.max(np.abs(A - B))
+
+
+class _Phi(object):
+    """Util class to have phi stft as callable without using
+    a lambda that does not pickle"""
+    def __init__(self, wsize, tstep, n_coefs):
+        self.wsize = wsize
+        self.tstep = tstep
+        self.n_coefs = n_coefs
+
+    def __call__(self, x):
+        return stft(x, self.wsize, self.tstep,
+                    verbose=False).reshape(-1, self.n_coefs)
+
+
+class _PhiT(object):
+    """Util class to have phi.T istft as callable without using
+    a lambda that does not pickle"""
+    def __init__(self, tstep, n_freq, n_step, n_times):
+        self.tstep = tstep
+        self.n_freq = n_freq
+        self.n_step = n_step
+        self.n_times = n_times
+
+    def __call__(self, z):
+        return istft(z.reshape(-1, self.n_freq, self.n_step), self.tstep,
+                     self.n_times)
+
+
+@verbose
+def tf_mixed_norm_solver(M, G, alpha_space, alpha_time, wsize=64, tstep=4,
+                         n_orient=1, maxit=200, tol=1e-8, log_objective=True,
+                         lipschitz_constant=None, debias=True, verbose=None):
+    """Solves TF L21+L1 inverse solver
+
+    Algorithm is detailed in:
+
+    A. Gramfort, D. Strohmeier, J. Haueisen, M. Hamalainen, M. Kowalski
+    Time-Frequency Mixed-Norm Estimates: Sparse M/EEG imaging with
+    non-stationary source activations
+    Neuroimage, in press as of Dec 2012
+
+    Functional Brain Imaging with M/EEG Using Structured Sparsity in
+    Time-Frequency Dictionaries
+    Gramfort A., Strohmeier D., Haueisen J., Hamalainen M. and Kowalski M.
+    INFORMATION PROCESSING IN MEDICAL IMAGING
+    Lecture Notes in Computer Science, 2011, Volume 6801/2011,
+    600-611, DOI: 10.1007/978-3-642-22092-0_49
+    http://dx.doi.org/10.1007/978-3-642-22092-0_49
+
+    Parameters
+    ----------
+    M : array
+        The data.
+    G : array
+        The forward operator.
+    alpha_space : float
+        The spatial regularization parameter. It should be between 0 and 100.
+    alpha_time : float
+        The temporal regularization parameter. The higher it is the smoother
+        will be the estimated time series.
+    wsize: int
+        length of the STFT window in samples (must be a multiple of 4).
+    tstep: int
+        step between successive windows in samples (must be a multiple of 2,
+        a divider of wsize and smaller than wsize/2) (default: wsize/2).
+    n_orient : int
+        The number of orientation (1 : fixed or 3 : free or loose).
+    maxit : int
+        The number of iterations.
+    tol : float
+        If absolute difference between estimates at 2 successive iterations
+        is lower than tol, the convergence is reached.
+    log_objective : bool
+        If True, the value of the minimized objective function is computed
+        and stored at every iteration.
+    lipschitz_constant : float | None
+        The lipschitz constant of the spatio temporal linear operator.
+        If None it is estimated.
+    debias : bool
+        Debias source estimates.
+    verbose : bool, str, int, or None
+        If not None, override default verbose level (see mne.verbose).
+
+    Returns
+    -------
+    X : array
+        The source estimates.
+    active_set : array
+        The mask of active sources.
+    E : list
+        The value of the objective function at each iteration. If log_objective
+        is False, it will be empty.
+    """
+    n_sensors, n_times = M.shape
+    n_dipoles = G.shape[1]
+
+    n_step = int(ceil(n_times / float(tstep)))
+    n_freq = wsize / 2 + 1
+    n_coefs = n_step * n_freq
+    phi = _Phi(wsize, tstep, n_coefs)
+    phiT = _PhiT(tstep, n_freq, n_step, n_times)
+
+    Z = np.zeros((0, n_coefs), dtype=np.complex)
+    active_set = np.zeros(n_dipoles, dtype=np.bool)
+    R = M.copy()  # residual
+
+    if lipschitz_constant is None:
+        lipschitz_constant = 1.1 * tf_lipschitz_constant(M, G, phi, phiT)
+
+    logger.info("lipschitz_constant : %s" % lipschitz_constant)
+
+    t = 1.0
+    Y = np.zeros((n_dipoles, n_coefs), dtype=np.complex)  # FISTA aux variable
+    Y[active_set] = Z
+    E = []  # track cost function
+    Y_time_as = None
+    Y_as = None
+
+    alpha_time_lc = alpha_time / lipschitz_constant
+    alpha_space_lc = alpha_space / lipschitz_constant
+    for i in xrange(maxit):
+        Z0, active_set_0 = Z, active_set  # store previous values
+
+        if active_set.sum() < len(R) and Y_time_as is not None:
+            # trick when using tight frame to do a first screen based on
+            # L21 prox (L21 norms are not changed by phi)
+            GTR = np.dot(G.T, R) / lipschitz_constant
+            A = GTR.copy()
+            A[Y_as] += Y_time_as
+            _, active_set_l21 = prox_l21(A, alpha_space_lc, n_orient)
+            # just compute prox_l1 on rows that won't be zeroed by prox_l21
+            B = Y[active_set_l21] + phi(GTR[active_set_l21])
+            Z, active_set_l1 = prox_l1(B, alpha_time_lc, n_orient)
+            active_set_l21[active_set_l21] = active_set_l1
+            active_set_l1 = active_set_l21
+        else:
+            Y += np.dot(G.T, phi(R)) / lipschitz_constant  # ISTA step
+            Z, active_set_l1 = prox_l1(Y, alpha_time_lc, n_orient)
+
+        Z, active_set_l21 = prox_l21(Z, alpha_space_lc, n_orient,
+                                shape=(-1, n_freq, n_step), is_stft=True)
+        active_set = active_set_l1
+        active_set[active_set_l1] = active_set_l21
+
+        # Check convergence : max(abs(Z - Z0)) < tol
+        stop = (safe_max_abs(Z, True - active_set_0[active_set]) < tol and
+                safe_max_abs(Z0, True - active_set[active_set_0]) < tol and
+                safe_max_abs_diff(Z, active_set_0[active_set],
+                                  Z0, active_set[active_set_0]) < tol)
+        if stop:
+            print 'Convergence reached !'
+            break
+
+        # FISTA 2 steps
+        # compute efficiently : Y = Z + ((t0 - 1.0) / t) * (Z - Z0)
+        t0 = t
+        t = 0.5 * (1.0 + sqrt(1.0 + 4.0 * t ** 2))
+        Y.fill(0.0)
+        dt = ((t0 - 1.0) / t)
+        Y[active_set] = (1.0 + dt) * Z
+        if len(Z0):
+            Y[active_set_0] -= dt * Z0
+        Y_as = active_set_0 | active_set
+
+        Y_time_as = phiT(Y[Y_as])
+        R = M - np.dot(G[:, Y_as], Y_time_as)
+
+        if log_objective:  # log cost function value
+            Z2 = np.abs(Z)
+            Z2 **= 2
+            X = phiT(Z)
+            RZ = M - np.dot(G[:, active_set], X)
+            pobj = 0.5 * linalg.norm(RZ, ord='fro') ** 2 \
+               + alpha_space * norm_l21(X, n_orient) \
+               + alpha_time * np.sqrt(np.sum(Z2.T.reshape(-1, n_orient),
+                                             axis=1)).sum()
+            E.append(pobj)
+            logger.info("Iteration %d :: pobj %f :: n_active %d" % (i + 1,
+                        pobj, np.sum(active_set)))
+        else:
+            logger.info("Iteration %d" % i + 1)
+
+    X = phiT(Z)
 
     if (active_set.sum() > 0) and debias:
         bias = compute_bias(M, G[:, active_set], X, n_orient=n_orient)